+++ app/persistance/graphdb/queries/construct_all_from_graph.sparql
