+++ app/persistance/graphdb/queries/construct_all_from_versioned_graph.sparql
