+++ app/persistance/graphdb/queries/count_triples.sparql
