+++ app/persistance/graphdb/queries/delta_query_deletions.sparql
