+++ app/persistance/graphdb/queries/delta_query_insertions.sparql
