+++ app/persistance/graphdb/queries/drop_graph.sparql
