+++ app/persistance/graphdb/queries/query_all_from_graph.sparql
