+++ app/persistance/graphdb/queries/query_creation_timestamps.sparql
