+++ app/persistance/graphdb/queries/query_snapshot_classes.sparql
