+++ app/persistance/graphdb/queries/query_snapshot_properties.sparql
