+++ app/persistance/graphdb/queries/query_static_core_triples.sparql
