+++ app/persistance/graphdb/queries/query_version_oblivious_triples.sparql
