+++ app/services/PollingTask.py
@@ -7,11 +7,11 @@ import requests
 from typing import Callable
 
 
-from app.Database import Session, engine
+from app.persistance.Database import Session, engine
 from app.models.DatasetModel import Dataset
 from app.models.TrackingTaskModel import TrackingTaskDto
 from app.services.VersioningService import StarVersService
-from app.utils.graphdb.GraphDatabaseUtils import create_engine
+from app.persistance.graphdb.GraphDatabaseUtils import create_engine
 from app.LoggingConfig import get_logger
 from app.services.MetricsService import MetricsService
