+++ docker-dev/docker-compose.dev.yml
@@ -91,7 +91,7 @@ services:
       - ${data_storage}/evaluation:/data/evaluation
       - ${data_storage}/logs:/data/logs
       - ${data_storage}/graphdb-data/graphdb-import:/graphdb-import
-    entrypoint: ["python", "evaluation/Compute.py", "orkg", "sm"]
+    entrypoint: ["python", "evaluation/compute.py", "orkg", "sm"]
   
   # Creates plots from the timing files
   create_plots:
