+++ docker-prod/docker-compose.prod.yml
@@ -87,15 +87,15 @@ services:
       - starvers_prod_net
 
   # Runs versioning and metrics computation tasks for data in ${data_storage}/evaluation. Possible executions are
-  # * entrypoint: ["python", "app/utils/Compute.py", "<repo_name>"]
+  # * entrypoint: ["python", "app/utils/compute.py", "<repo_name>"]
   #   Runs versioning and metrics computation for the specified repository
-  # * entrypoint: ["python", "app/utils/Compute.py", "<repo_name>", from_scratch, "yyyyMMdd-HHmmss_sss"]
+  # * entrypoint: ["python", "app/utils/compute.py", "<repo_name>", from_scratch, "yyyyMMdd-HHmmss_sss"]
   #   Runs versioning and metrics computation for the specified repository from a specific timestamp with initial versioning
-  # * entrypoint: ["python", "app/utils/Compute.py", "<repo_name>", from_version, "yyyyMMdd-HHmmss_sss"]
+  # * entrypoint: ["python", "app/utils/compute.py", "<repo_name>", from_version, "yyyyMMdd-HHmmss_sss"]
   #   Continues versioning and metrics computation for the specified repository from a specific timestamp without initial versioning
-  # * entrypoint: ["python", "app/utils/Compute.py", "<repo_name>",  (<from_version|from_scratch>), ("yyyyMMdd-HHmmss_sss"), "v"]
+  # * entrypoint: ["python", "app/utils/compute.py", "<repo_name>",  (<from_version|from_scratch>), ("yyyyMMdd-HHmmss_sss"), "v"]
   #   Runs only versioning without metrics computation for the specified repository from a specific timestamp
-  # * entrypoint: ["python", "app/utils/Compute.py", "<repo_name>", (<from_version|from_scratch>), ("yyyyMMdd-HHmmss_sss"), "dm, sm"]
+  # * entrypoint: ["python", "app/utils/compute.py", "<repo_name>", (<from_version|from_scratch>), ("yyyyMMdd-HHmmss_sss"), "dm, sm"]
   #   Runs only thout metrics computation without versioning for the specified repository, with the possibility of running from scratch or from a specific timestamp
   compute:
     image: starversserver
@@ -112,7 +112,7 @@ services:
       - ${data_storage}/evaluation:/data/evaluation
       - ${data_storage}/logs:/data/logs
       - ${data_storage}/graphdb-data/graphdb-import:/graphdb-import
-    entrypoint: ["python", "evaluation/Compute.py", "orkg", "sm", "from_version", "20250718-175230_258"] 
+    entrypoint: ["python", "evaluation/compute.py", "orkg", "sm", "from_version", "20250828-124726_863"] 
 
     # Creates plots from the timing files
   create_plots:
